refactor(wework): split desktop workbench layout
