+++ wework/src/components/layout/useResizableSidebar.ts
@@ -1,0 +1,61 @@
+import { useState, type PointerEvent } from 'react'
+
+const DEFAULT_SIDEBAR_WIDTH = 280
+const MIN_SIDEBAR_WIDTH = 220
+const MAX_SIDEBAR_WIDTH = 420
+const SIDEBAR_WIDTH_STORAGE_KEY = 'wework.desktop.sidebar.width'
+
+function clampSidebarWidth(width: number): number {
+  return Math.min(MAX_SIDEBAR_WIDTH, Math.max(MIN_SIDEBAR_WIDTH, width))
+}
+
+function readStoredSidebarWidth(): number {
+  if (typeof window === 'undefined') return DEFAULT_SIDEBAR_WIDTH
+
+  const storedWidth = window.localStorage.getItem(SIDEBAR_WIDTH_STORAGE_KEY)
+  const parsedWidth = storedWidth ? Number(storedWidth) : DEFAULT_SIDEBAR_WIDTH
+
+  if (!Number.isFinite(parsedWidth)) {
+    return DEFAULT_SIDEBAR_WIDTH
+  }
+
+  return clampSidebarWidth(parsedWidth)
+}
+
+function storeSidebarWidth(width: number) {
+  window.localStorage.setItem(SIDEBAR_WIDTH_STORAGE_KEY, String(width))
+}
+
+export function useResizableSidebar() {
+  const [sidebarWidth, setSidebarWidth] = useState(readStoredSidebarWidth)
+
+  const handleResizeStart = (event: PointerEvent<HTMLButtonElement>) => {
+    event.preventDefault()
+    const previousCursor = document.body.style.cursor
+    const previousUserSelect = document.body.style.userSelect
+
+    document.body.style.cursor = 'col-resize'
+    document.body.style.userSelect = 'none'
+
+    const handlePointerMove = (moveEvent: globalThis.PointerEvent) => {
+      const nextWidth = clampSidebarWidth(moveEvent.clientX)
+      setSidebarWidth(nextWidth)
+      storeSidebarWidth(nextWidth)
+    }
+
+    const handlePointerUp = () => {
+      document.body.style.cursor = previousCursor
+      document.body.style.userSelect = previousUserSelect
+      document.removeEventListener('pointermove', handlePointerMove)
+      document.removeEventListener('pointerup', handlePointerUp)
+    }
+
+    document.addEventListener('pointermove', handlePointerMove)
+    document.addEventListener('pointerup', handlePointerUp)
+  }
+
+  return {
+    sidebarWidth,
+    handleResizeStart,
+  }
+}
